refactor: move TradingEngineSnapshotSummary to a new file
